Added models Comment, Category and Tag.
I also added Tag because it took me 1 minute. Please note that the model Comment
does not need a field to know if the user is the owner of the product. This can
be calculated at runtime, so I didn't include that field.

I also added some small details, such as the icon and description for the
categories.

This is also my first commit, let's see how it goes :)