Added getter/setter methods for Tag, Category and Comment.
Also added blank=False to required fields and set defaults for
some of the fields. Added a default icon for an unknown category and created the
directory web/icons where to place icons for categories.

In addition to this, the value -1 in parent_id shall be used to state that the
element has no parent (e.g. in comments or categories).