Added db.sql.
Guys, you have to add db.sql, where all the data is stored. In order to do this,
you should edit your settings.py and have something like this:

DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.sqlite3', # Add 'postgresql_psycopg2', 'postgresql', 'mysql', 'sqlite3' or 'oracle'.
        'NAME': 'db.sql',                      # Or path to database file if using sqlite3.
        'USER': '',                      # Not used with sqlite3.
        'PASSWORD': '',                  # Not used with sqlite3.
        '': '',                      # Set to empty string for localhost. Not used with sqlite3.
        'PORT': '',                      # Set to empty string for default. Not used with sqlite3.
    }
}

After editing settings.py, just run 'python manage.py syncdb' and it will create
the database for you. It will ask you to have a superuser for the database, so
just create an account and that's it.