Basic user log-in and session.
The sign up form now works, although it needs huge improvements. The user
password is converted to sha1 without salt, and then it is compared with the
password in the log-in form.

In order to keep sessions, a cookie is set using 'id' with the user ID that is
logged in. WARNING: every view in views.py should check if there is an on-going
session and act accordingly.

Missing things: form validation, editing own profile, adding a picture, creating
custom pages for common errors (e.g. username must be unique, and Django
provides its own 'ugly' error page).